refactor: #455 move Project to utils directory
For reasons similar to the ones of previous commit.
