+++ packages/generator-liferay-theme/lib/utils/Project.js
